pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "LLM-Agent-with-docker"
version = "0.1.0"
description = "This project implements a Postgres Agent that runs code in a docker container."
readme = "README.md"
requires-python = ">=3.7"

dependencies = [
    "aiohappyeyeballs==2.3.5",
    "aiohttp==3.10.1",
    "aiosignal==1.3.1",
    "annotated-types==0.7.0",
    "anyio==4.4.0",
    "attrs==24.2.0",
    "certifi==2024.7.4",
    "charset-normalizer==3.3.2",
    "contourpy==1.2.1",
    "cycler==0.12.1",
    "dataclasses-json==0.6.7",
    "distro==1.9.0",
    "docker==7.1.0",
    "fonttools==4.53.1",
    "frozenlist==1.4.1",
    "greenlet==3.0.3",
    "h11==0.14.0",
    "httpcore==1.0.5",
    "httpx==0.27.0",
    "idna==3.7",
    "jiter==0.5.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "kiwisolver==1.4.5",
    "langchain==0.2.12",
    "langchain-community==0.2.11",
    "langchain-core==0.2.29",
    "langchain-openai==0.1.20",
    "langchain-text-splitters==0.2.2",
    "langgraph==0.2.1",
    "langgraph-checkpoint==1.0.2",
    "langsmith==0.1.98",
    "marshmallow==3.21.3",
    "matplotlib==3.9.1.post1",
    "multidict==6.0.5",
    "mypy-extensions==1.0.0",
    "numpy==1.26.4",
    "openai==1.40.1",
    "orjson==3.10.6",
    "packaging==24.1",
    "pandas==2.2.2",
    "pillow==10.4.0",
    "psycopg2==2.9.9",
    "pydantic==2.8.2",
    "pydantic_core==2.20.1",
    "pyparsing==3.1.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "pytz==2024.1",
    "PyYAML==6.0.2",
    "regex==2024.7.24",
    "requests==2.32.3",
    "six==1.16.0",
    "sniffio==1.3.1",
    "SQLAlchemy==2.0.32",
    "tenacity==8.5.0",
    "tiktoken==0.7.0",
    "tqdm==4.66.5",
    "typing-inspect==0.9.0",
    "typing_extensions==4.12.2",
    "tzdata==2024.1",
    "urllib3==2.2.2",
    "yarl==1.9.4",
]

[project.optional-dependencies]
dev = ["black", "flake8"]
test = ["pytest", "pytest-cov"]

[tool.black]
line-length = 88
target-version = ['py37']

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--strict-markers"
